Add generateFood function to streamline food creation logic
